+++ src/components/layouts/About/OurTeam/ODC.tsx
@@ -16,8 +16,6 @@ const ODC = () => {
     ...data.allStaffs,
   };
 
-  console.log("")
-
   return (
     <Box>
       <Fragment>
