+++ src/components/layouts/Blog/Blog.tsx
@@ -8,21 +8,97 @@ import {
   Heading,
   Image,
 } from "@chakra-ui/react";
-import React, { useState } from "react";
-import { BLOGS, CATEGORIES } from "data/blog";
+import React, { useEffect, useState } from "react";
+import {CATEGORIES } from "data/blog";
 import { useRouter } from "next/router";
 import MobileBlogMenu from "./MobileBlogMenu";
+import Editable from "components/organisms/Editable/Editable";
+import {
+  combinedConfig,
+  combinedConstant,
+} from "config/constants/editable-copy/combined";
+import { useCopyData } from "contexts/EditableCopyContext";
 
-const Blog = () => {
-  const blogs = BLOGS;
+export const blogCategories = {
+  news: "News & Events",
+  recipe: "Recipes",
+  testimonial: "Testimonials",
+  hospitality: "Hospitality Articles",
+  others: "Others",
+};
 
-  const [currentBlog, setCurrentBlog] = useState(blogs[0]);
+export type BlogCategory = keyof typeof blogCategories;
+
+export type BlogPost = {
+  id: string;
+  name?: string;
+  text?: string;
+  isTopPost: boolean;
+  postDate: Date;
+  postsImg: string;
+  coverImage?: string;
+  postHeading: string;
+  postsText: string;
+};
+
+export type Blogs = {
+  category: BlogCategory;
+  posts: BlogPost[];
+};
+
+type BlogProps = {
+  currentBlog: Blogs;
+  setCurrentBlog: (value: Blogs) => void;
+};
+
+export const formatDate = (dateObj: Date) => {
+  const months = [
+    "January",
+    "February",
+    "March",
+    "April",
+    "May",
+    "June",
+    "July",
+    "August",
+    "September",
+    "October",
+    "November",
+    "December",
+  ];
+
+  const day = dateObj.getDate();
+  const month = dateObj.getMonth();
+  const year = dateObj.getFullYear();
+  return `${months[month]} ${day}, ${year}`;
+};
+
+const Blog = (props: BlogProps) => {
+  const { currentBlog, setCurrentBlog } = props;
+
+  const [selectedCategoryPosts, setSelectedCategoryPosts] = useState<
+    BlogPost[]
+  >([]);
+
+  useEffect(() => {
+    setSelectedCategoryPosts(currentBlog.posts);
+  }, [currentBlog.posts]);
+
+  const { blogConfig } = combinedConfig;
+
+  const { data } = useCopyData();
+
+  const { BLOGS } = {
+    ...combinedConstant.blog,
+    ...data.blog,
+  };
+
   const categories = CATEGORIES;
 
   const router = useRouter();
 
-  const ShowPosts = (post: any) => {
-    router.push(`/posts/${post.postId}`);
+  const ShowPosts = (post: BlogPost) => {
+    router.push(`/posts/${post.id}`);
   };
 
   return (
@@ -78,101 +154,141 @@ const Blog = () => {
             currentBlog={currentBlog}
             setCurrentBlog={setCurrentBlog}
             categories={categories}
-            blogs={blogs}
+            blogs={BLOGS}
             ShowPosts={ShowPosts}
           />
         </Flex>
       </Box>
 
-      <Box
-        w={{
-          xl: "830px",
-          "2xl": "1000px",
-        }}
+      <Editable
+        defaultValues={selectedCategoryPosts as any}
+        config={blogConfig.blog}
+        page="blog"
       >
-        <Grid
-          gridTemplateColumns={{
-            sm: "auto",
-            md: "1fr 1fr",
-            lg: "1fr 1fr",
-            xl: "1fr 1fr",
-            "2xl": "1fr 1fr",
+        <Box
+          w={{
+            xl: "830px",
+            "2xl": "1000px",
           }}
-          w="100%"
-          cursor="pointer"
         >
-          {currentBlog?.posts?.map((post, index) => {
-            return (
-              <Box
-                key={index}
-                bg={{
-                  sm: post.bg,
-                  md: "none",
-                  lg: "none",
-                  xl: "none",
-                  "2xl": "none",
-                }}
-                mx={{
-                  sm: "0",
-                  md: "0px",
-                  lg: "0px",
-                  xl: "20px",
-                  "2xl": "20px",
-                }}
-                p={{
-                  sm: "20px",
-                  md: "30px",
-                  lg: "40px",
-                  xl: "0",
-                  "2xl": "0",
-                }}
-                mb="50px"
-                onClick={() => ShowPosts(post)}
-                transition="all ease 0.5s"
-                _hover={{
-                  transform: "scale(1.03)",
-                }}
-              >
-                <Image
+          <Grid
+            gridTemplateColumns={{
+              sm: "auto",
+              md: "1fr 1fr",
+              lg: "1fr 1fr",
+              xl: "1fr 1fr",
+              "2xl": "1fr 1fr",
+            }}
+            w="100%"
+            cursor="pointer"
+          >
+            {currentBlog?.posts?.map((post, index) => {
+              return (
+                <Box
                   key={index}
-                  alt="blog-img"
-                  src={post.postsImg}
-                  borderRadius="3px"
-                  mx="auto"
-                  w="100%"
-                  mb={{
+                  bg={{
+                    sm: post.postsImg,
+                    md: "none",
+                    lg: "none",
+                    xl: "none",
+                    "2xl": "none",
+                  }}
+                  mx={{
+                    sm: "0",
+                    md: "0px",
+                    lg: "0px",
+                    xl: "20px",
+                    "2xl": "20px",
+                  }}
+                  p={{
                     sm: "20px",
-                    md: "20px",
-                    lg: "20px",
+                    md: "30px",
+                    lg: "40px",
                     xl: "0",
                     "2xl": "0",
                   }}
-                />
-
-                <Text
-                  mt="8px"
-                  color="rgba(2, 29, 55, 0.44)"
-                  fontFamily="Manrope"
-                  fontWeight="400"
-                  fontSize="18px"
-                  lineHeight="25px"
-                  display={{
-                    sm: "none",
-                    md: "none",
-                    lg: "none",
-                    xl: "block",
-                    "2xl": "block",
+                  mb="50px"
+                  onClick={() => ShowPosts(post)}
+                  transition="all ease 0.5s"
+                  _hover={{
+                    transform: "scale(1.03)",
                   }}
                 >
-                  {post.postsDate}
-                </Text>
+                  <Image
+                    key={index}
+                    alt="blog-img"
+                    src={post.postsImg}
+                    borderRadius="3px"
+                    mx="auto"
+                    w="100%"
+                    mb={{
+                      sm: "20px",
+                      md: "20px",
+                      lg: "20px",
+                      xl: "0",
+                      "2xl": "0",
+                    }}
+                  />
 
-                <Box>
-                  <Box
-                    w="88px"
-                    h="3px"
-                    bg="#021D37"
-                    mb="10px"
+                  <Text
+                    mt="8px"
+                    color="rgba(2, 29, 55, 0.44)"
+                    fontFamily="Manrope"
+                    fontWeight="400"
+                    fontSize="18px"
+                    lineHeight="25px"
+                    display={{
+                      sm: "none",
+                      md: "none",
+                      lg: "none",
+                      xl: "block",
+                      "2xl": "block",
+                    }}
+                  >
+                    {formatDate(post.postDate)}
+                  </Text>
+
+                  <Box>
+                    <Box
+                      w="88px"
+                      h="3px"
+                      bg="#021D37"
+                      mb="10px"
+                      display={{
+                        sm: "block",
+                        md: "block",
+                        lg: "block",
+                        xl: "none",
+                        "2xl": "none",
+                      }}
+                    ></Box>
+
+                    <Heading
+                      color="#021d37"
+                      fontFamily="Playfair Display"
+                      fontWeight="700"
+                      fontSize={{
+                        sm: "20px",
+                        md: "22px",
+                        lg: "22px",
+                        xl: "22px",
+                        "2xl": "22px",
+                      }}
+                      textTransform="capitalize"
+                      lineHeight="30px"
+                    >
+                      {post.postHeading}
+                    </Heading>
+                  </Box>
+
+                  <Text
+                    mt="4px"
+                    mb="20px"
+                    color="rgba(2, 29, 55, 0.44)"
+                    fontFamily="Manrope"
+                    fontWeight="400"
+                    fontSize="18px"
+                    lineHeight="25px"
                     display={{
                       sm: "block",
                       md: "block",
@@ -180,83 +296,49 @@ const Blog = () => {
                       xl: "none",
                       "2xl": "none",
                     }}
-                  ></Box>
-
-                  <Heading
-                    color="#021d37"
-                    fontFamily="Playfair Display"
-                    fontWeight="700"
-                    fontSize={{
-                      sm: "20px",
-                      md: "22px",
-                      lg: "22px",
-                      xl: "22px",
-                      "2xl": "22px",
-                    }}
-                    textTransform="capitalize"
-                    lineHeight="30px"
                   >
-                    {post.postHeading}
-                  </Heading>
+                    {formatDate(post.postDate)}
+                  </Text>
+
+                  <Text
+                    mt="10px"
+                    color="#021d37"
+                    fontFamily="Manrope"
+                    fontWeight="400"
+                    fontSize="18px"
+                    lineHeight="25px"
+                    noOfLines={5}
+                    dangerouslySetInnerHTML={{ __html: post.postsText }}
+                  />
+                  <Button
+                    w="142px"
+                    height="46.89px"
+                    border="2px solid #021D37"
+                    bg="transparent"
+                    fontWeight="700"
+                    fontSize="16px"
+                    lineHeight="22px"
+                    textAlign="center"
+                    borderRadius="3px"
+                    mt="25px"
+                    onClick={() => ShowPosts(post)}
+                    transition="all ease 0.5s"
+                    display={{
+                      sm: "block",
+                      md: "block",
+                      lg: "block",
+                      xl: "none",
+                      "2xl": "none",
+                    }}
+                  >
+                    LEARN MORE
+                  </Button>
                 </Box>
-
-                <Text
-                  mt="4px"
-                  mb="20px"
-                  color="rgba(2, 29, 55, 0.44)"
-                  fontFamily="Manrope"
-                  fontWeight="400"
-                  fontSize="18px"
-                  lineHeight="25px"
-                  display={{
-                    sm: "block",
-                    md: "block",
-                    lg: "block",
-                    xl: "none",
-                    "2xl": "none",
-                  }}
-                >
-                  {post.postsDate}
-                </Text>
-
-                <Text
-                  mt="10px"
-                  color="#021d37"
-                  fontFamily="Manrope"
-                  fontWeight="400"
-                  fontSize="18px"
-                  lineHeight="25px"
-                  noOfLines={5}
-                  dangerouslySetInnerHTML={{ __html: post.postsText }}
-                />
-                <Button
-                  w="142px"
-                  height="46.89px"
-                  border="2px solid #021D37"
-                  bg="transparent"
-                  fontWeight="700"
-                  fontSize="16px"
-                  lineHeight="22px"
-                  textAlign="center"
-                  borderRadius="3px"
-                  mt="25px"
-                  onClick={(e) => ShowPosts(e)}
-                  transition="all ease 0.5s"
-                  display={{
-                    sm: "block",
-                    md: "block",
-                    lg: "block",
-                    xl: "none",
-                    "2xl": "none",
-                  }}
-                >
-                  LEARN MORE
-                </Button>
-              </Box>
-            );
-          })}
-        </Grid>
-      </Box>
+              );
+            })}
+          </Grid>
+        </Box>
+      </Editable>
     </Flex>
   );
 };
