+++ src/config/constants/editable-copy/blog/values.ts
@@ -1,0 +1,524 @@
+import { Blogs } from "components/layouts/Blog/Blog";
+import { v4 as uuidv4 } from "uuid";
+
+const blogPost = "/assets/imgs/blogpost.png";
+const blogPost2 = "/assets/imgs/aboutus.png";
+const suicideEvent = "/assets/imgs/suicide-event.jpeg";
+const uchyfide = "/assets/imgs/uchyfide.jpeg";
+const Udoka = "/assets/imgs/Udoka.jpeg";
+const NBTEdirector = "/assets/imgs/NBTEdirector.jpg";
+const nbte = "/assets/imgs/nbte.jpeg";
+const cookFish = "/assets/imgs/cookFish.jpg";
+const catfish = "/assets/imgs/catfish.png";
+const hotel = "/assets/imgs/hotel.jpg";
+const decorating = "/assets/imgs/decorating.jpg";
+const airport = "/assets/imgs/AIRPORT.jpg";
+const airTravels = "/assets/imgs/air-travels.jpg";
+const ekpangNkukwo = "/assets/imgs/Ekpang-Nkukwo.jpg";
+const studentLetter = "/assets/imgs/student-letter.png";
+const developingSkills = "/assets/imgs/developing-skills.jpg";
+const finishingStrong = "/assets/imgs/finishing-strong.png";
+const foodFair1 = "/assets/imgs/food_fair_1.jpg";
+const foodFair2 = "/assets/imgs/food_fair_2.jpg";
+const careerDrive1 = "/assets/imgs/career_drive1.jpg";
+const careerDrive2 = "/assets/imgs/career_drive2.jpg";
+const instructorTraining = "/assets/imgs/instructors_training.jpg";
+const instructorTraining2 = "/assets/imgs/instructors_training2.jpg";
+const lsgSkyChef = "/assets/imgs/lsg-_sky_chef_1.jpg";
+const lsgSkyChef2 = "/assets/imgs/lsg-_sky_chef_2.jpg";
+const wheatbaker = "/assets/imgs/wheatbaker.jpg";
+const wheatbaker2 = "/assets/imgs/wheatbaker2.jpg";
+const arise = "/assets/imgs/arise.jpg";
+const arise2 = "/assets/imgs/arise2.jpg";
+const ishkTolaram = "/assets/imgs/ishkTolaram.jpg";
+const ishkTolaram2 = "/assets/imgs/ishkTolaram2.jpg";
+const internship = "/assets/imgs/internship.jpg";
+const swiss = "/assets/imgs/swiss.jpg";
+const swiss2 = "/assets/imgs/swiss2.jpg";
+const staffSeminar = "/assets/imgs/staff-seminar.jpg";
+const staffSeminar2 = "/assets/imgs/staff-seminar-2.jpg";
+
+export const BLOGS: Blogs[] = [
+  {
+    category: "news",
+    posts: [
+      {
+        id: uuidv4(),
+        isTopPost: true,
+        postsImg: finishingStrong,
+        postDate: new Date("Wed Jun 15 2022 00:00:00 GMT+0100"),
+        postHeading:
+          "Wavecrest College received the SKILLS INITIATIVE FOR AFRICA(SIFA) grant for the construction of the WAVECREST-HSEDCE",
+        postsText: `Through this expansion, we will train 1,000 women annually and serve the growing demands of skilled labor in the Hospitality Industry`,
+      },
+      {
+        id: uuidv4(),
+        postDate: new Date("Wed Jun 15 2022 00:00:00 GMT+0100"),
+        isTopPost: true,
+        postsImg: foodFair1,
+        postHeading: "fundraising food fair",
+        coverImage: foodFair2,
+        postsText: `The Food Fair was held on the 24th of June, 2023, at the Wavecrest College.
+
+                THE OBJECTIVES
+
+                It was a fundraising and publicity event to raise funds for our counterpart contribution of the SIFA (Skills Initiative For Africa) Grant received for the construction of the Wavecrest Hospitality Skills and Enterprise Development Centre of Excellence  - HSEDCE.
+
+                ALL HANDS ON DECK
+
+                All departments of the College participated actively in collaboration with the students. Admin Department was in charge of the general coordination, the ODC (Outdoor Catering) and Student Assistant gave their support and assisted with purchasing the items, foodstuff, and materials needed. The Accounts department was in charge of the collection of funds and the F & B Lecturers supervised and coordinated the students.
+                We also had external support from Chef Cuddles (Yejide Uthman), Sugar Creations, and Mrs. Kelechi Nwankwo who served as volunteers.
+
+                The turnout was good as over 120 students from different secondary schools were in attendance as well as friends, family members and food enthusiasts who came to support and create a vibrant atmosphere.
+
+                The Food Fair was a great experience`,
+      },
+      {
+        id: uuidv4(),
+        postDate: new Date("Wed Jun 15 2022 00:00:00 GMT+0100"),
+        isTopPost: true,
+        postsImg: ishkTolaram2,
+        coverImage: ishkTolaram,
+        postHeading: "ISHK TOLARAM GRANT",
+        postsText: `Bridging the Hospitality Skills Gap.
+
+              The College received a grant from Ishk Tolaram Foundation to support Thirty (30) underserved women in the One Year Certificate In Hospitality Operations Programme.
+
+              The program aims to bridge the community gaps of access to international standard vocational training that gives the students hands-on experience and real-time practice, and exposure to state-of-the-art equipment which they will necessarily find in the industry.
+
+              The European Journal of Business and Management Research (EJBMR) - Vol. 7 No. 1 (2022) with the article titled Gap of Entrepreneurship training  shows the result of studies done to highlight the lack of skilled labour to fill the many job opportunities in this Hospitality Industry.
+
+              Training the Next Generation of Hospitality Professionals
+
+              Instilling our core values is high priority. Core values are an essential part of the curriculum- reflected in the acronym SPIRIT!  - Spirit of Service, Personal Responsibility, Integrity , Respect for others, Industriousness and Teamwork.`,
+      },
+      {
+        id: uuidv4(),
+        postDate: new Date("Wed Jun 15 2022 00:00:00 GMT+0100"),
+        isTopPost: true,
+        postsImg: instructorTraining,
+        coverImage: instructorTraining2,
+        postHeading: "Instructors Training Workshop ",
+        postsText: `Objective
+
+              The Instructor Training Workshop and Refresher Course at Wavecrest College was organized as a continuous professional development for educators within all  Hospitality Colleges, Schools, and Centres owned by of Women’s Board - Educational Corporation Society and their counterparts  Nigerian Association for Women’s Advancement (NAWA) and Wetland.
+
+              Institutions in Attendance
+
+              There were participants from Lantana College of Hospitality, Enugu, Orisun School of Hospitality, Ibadan, Iroto School of Hotel and Catering, Ogun State, Lagoon Institute of Hospitality Studies, Lagos, and Wavecrest College of Hospitality.
+
+              Training Content
+
+              The Two-Day training workshop covered a wide range of activities, including the demonstration of different beef dishes and presentations, introduction to new teaching methodologies and technologies, exposure to new and advanced equipment and the Demonstration of rational Ovens -   Rational combi steamer & Rational 1 combi Pro
+
+              Hands-on Learning
+
+              The facilitator for the day was Chef Mayowa Olatunji who delivered excellently well.
+
+              The training concluded with an Industrial visit to BCE Food Service Equipment, Matori, Nigeria, one of the leading companies in Africa for supply and delivery of all kinds of commercial catering equipment and products, kitchen utensils, industrial cookware & and appliances for industries in food service and hospitality.
+              `,
+      },
+      {
+        id: uuidv4(),
+        postDate: new Date("Wed Jun 15 2022 00:00:00 GMT+0100"),
+        isTopPost: false,
+        postsImg: internship,
+        coverImage: internship,
+        postHeading: "2023 INTERNSHIP IN TOUCH WITH THE INDUSTRY",
+        postsText: `CREATING A CULTURE OF EXCELLENCE
+
+              The Student Industrial Work Experience Scheme (SIWES) is an essential component of our academic program, providing students with practical exposure to the real world of work.
+
+              76 STUDENTS WERE PLACED ON INTERNSHIP IN THE FOLLOWING ORGANIZATIONS:
+
+               - 355 RESTAURANT Ikeja
+               - April Gossip Restaurant
+               - Cake n candy Ikeja
+               - Cakement Aguda
+               - Cakes by somcheZ
+               - Clares Cakes and Pastries
+               - Delta kitchen
+               - Doodle cake Festac
+               - Eko Hotel and Suites
+               - Four Points by Sheraton
+               - Hotel Ibis Hotel Ikeja
+               - Hotel Ibis Royale
+               - Ikoyi Club 1938
+               - IN4U Cakes and Events
+               - IN4U CAKES SURULERE
+               - Lagos Business School
+               - Lagos Continental Hotel VI
+               - Legend Hotel Lagos Airport, Curio Collection by Hilton
+               - Lounge 38
+               - LSG Sky Chefs Things Remembered
+               - Marriott International Ikeja
+               - Melbrooke Hotel
+               - Millenium Apartment Studio Lekki
+               - Montana Residence GRA Ikeja
+               - Mövenpick Hotel Ikoyi
+               - Newrest Nigeria
+               - OASIS HOTEL AGABARA
+               - Parkinn by Radisson
+               - Providence by Mantis Hotel Ikeja
+               - Raddisson Blu Hotel VI
+               - Royal Residence
+               - SEATTLE RESIDENCE VI
+               - Sugar creation
+               - SUN HEAVEN HOTEL VICTORIA ISLAND
+               - Suru Express Ikeja
+               - Suru Express Ikorodu
+               - The Art Hotel
+               - THE ART HOTEL VI
+               - The Cake House
+               - The George Hotel
+               - The Wheatbeaker
+               - Villa Monument Hotel Victoria Island
+               - Wavecrest College Outdoor Catering unit
+              `,
+      },
+      {
+        id: uuidv4(),
+        postDate: new Date("Wed Jun 15 2022 00:00:00 GMT+0100"),
+        isTopPost: false,
+        postsImg: staffSeminar,
+        coverImage: staffSeminar2,
+        postHeading: "STAFF WEEKEND WORKSHOP",
+        postsText: `
+              The 2023 staff weekend workshop proved to be a valuable and enriching experience for all participants.
+
+              Focus areas
+
+               - Unity of Purpose and Team building.
+               - Learning to Value Others.
+               - Our Christian identity
+               - Professionalism at work
+               - Financial Growth
+
+              Staff Engagement
+
+              It was a much-needed opportunity to unwind and reward our staff for the hard work and effort put into running the College this year.
+
+              Facilitators
+
+              Prof. Franca Ovadje - Founder, Executive Director, Danne Institute for Research,  Rev. Fr. Bonniface Okafor a Priest of Opus Dei, Mrs. Adejoke Ohiakena, Deputy Head of School, Corona School Trust and Nkemdilim Mojume - Lead Delivery Consultant at Franklin Covey West Africa
+
+              Outcome
+
+              The working sessions birthed SMART ‘way forward’ working document that would be our reference point for the academic session.
+              `,
+      },
+      {
+        id: uuidv4(),
+        postDate: new Date("Wed Jun 15 2022 00:00:00 GMT+0100"),
+        isTopPost: false,
+        postsImg: wheatbaker,
+        coverImage: wheatbaker2,
+        postHeading: "ONE-DAY INTERACTIVE SESSION AT WHEATBAKER HOTELS",
+        postsText: `
+              INDUSTRY CONNECTION
+
+              Students and members of staff were at The Wheatbaker - one of Nigeria's Leading Boutique Hotels. The invitation was extended to us by the General Manager Mr. Paul Kavanagh. This was proudly organized by a graduate of Wavecrest College who was then the Personal Assistant to the GM of The Wheatbaker in the person of Ms. Mercy Nwosu.
+
+              Hotel General Managers influencing  Hospitality Culture
+
+              The interactive session was with a panel of General Managers of leading hotels in Lagos - Four Points by Sheraton - Alex Richwell, Jara Beach Resort - Mr. Mark Slade, Ebonyi live, Radisson Blu - Mr. Ahmed Raza and George Residence Lagos - Morten Ebbesen to mention a few.
+
+              Learning Curves
+
+              The students were taught to have passion for the industry and not just to work for the money. Two of our students - Victory Ahamisi and Anthonia Orozele won a free spa session and a free lunch for two respectively for answering questions.
+
+              It was a well-organized and wonderful experience for our students and staff.
+              `,
+      },
+      {
+        id: uuidv4(),
+        postDate: new Date("Wed Jun 15 2022 00:00:00 GMT+0100"),
+        isTopPost: false,
+        postsImg: swiss,
+        coverImage: swiss2,
+        postHeading: "THE SWISS GOVERNMENT SCHOLARSHIP",
+        postsText: `Swiss Scholarship
+
+              In total, Thirty-eight (38) Learners/ Students have benefited from the scholarship in the last 6 years and each one has a unique story of transformation and achievement.
+              This once in-a-lifetime experience has been enriching for all participants and a great opportunity for professional development, adventures, learning opportunities, networking, cultural diversity and more.
+
+              Appreciation and Continuous Commitment
+
+              A lot more has been achieved with the Swiss Scholarship where our students have the opportunity to enrich their hospitality training and gain global exposure that further enhances their prospects in the Hospitality Job Market. Without your support, none of these success stories would have been told.
+              Thank you BIHC!
+              Thank you Rainbow Unlimited!
+              Thank you Swiss Government!
+
+              Recreating Lifestyles
+
+              Since 2017, Students from Wavecrest College of Hospitality, Lagos, Nigeria have been privileged recipients of the Swiss all-expense paid scholarship to study at the BOMA INTERNATIONAL HOTEL COLLEGE (BIHC), Kenya. BIHC is an affiliate of the Kenyan Red Cross in partnership with Business and Hotel Management School Switzerland.
+
+              2023 Cohort
+
+              The last batch of trainers arrived and returned on the 16th of December, 2023 fully energized and equipped to take up new challenges.
+
+              ‘Reflecting on my immersive experience at Boma International Hospitality College, I wish to unfold a narrative of profound experiences, learning, and personal growth. From the lively streets of Nairobi to the dynamic classrooms of Boma, this adventure has been nothing short of transformative.’ - Olivia Ofuasia
+              `,
+      },
+      {
+        id: uuidv4(),
+        postDate: new Date("Wed Jun 15 2022 00:00:00 GMT+0100"),
+        isTopPost: false,
+        postsImg: arise,
+        coverImage: arise2,
+        postHeading: "ARISE FOUNDATION CULINARY TRAINING",
+        postsText: `
+              Training Content
+
+              The Culinary and Entrepreneurship Skill Acquisition Program sponsored by Arise Foundation, was held between July 2022 and November 2023.
+
+              The program was aimed at providing culinary skills training to returned irregular migrants, women at risk of being trafficked due to economic conditions, and widows.
+
+              The training covered theoretical and basic practical hands-on culinary skills, covering National and Continental dishes, Yeast products, Pastries, Finger foods, Cake making and Icing techniques, Plating, and Food and Beverage table setting components.
+
+              Seminars & Mentoring
+
+              The orientation before the commencement of each Training Batch helped them gain an insight into what the training entailed.
+              The mentoring sessions and seminars provided a safe and supportive space for the participants. Hence they felt comfortable sharing their experiences without fear of judgment.
+
+              Impact
+
+              We completed five batches of culinary skills. Training a total of seventy-five (75) participants.
+              The final graduation ceremony took place on November 2023. In attendance were the Women’s Board and collaborators; Arise Representative, Franero Consultants/ Mdif-Migration and Development Impact Foundation.
+              `,
+      },
+      {
+        id: uuidv4(),
+        postDate: new Date("Wed Jun 15 2022 00:00:00 GMT+0100"),
+        isTopPost: false,
+        postsImg: lsgSkyChef,
+        coverImage: lsgSkyChef2,
+        postHeading: "LSG SKY CHEF FIELD TRIP",
+        postsText: `About LSG Sky Chef
+
+              LSG Skychefs is a leading global provider of culinary and hospitality solutions for the aviation industry.
+
+              They are located in Lagos, Nigeria and the facility has around 40 employees and produces approximately 450 Thousand meals per year.
+
+              Their services encompass a wide range of offerings, including menu design, food preparation, and logistics management to ensure that passengers enjoy a satisfying dining experience.
+
+              The Exposure
+
+              A total of 24 students and Lecturers were present at the field trip and they recounted how educational it was.
+
+              The Experience
+
+              The objective of the trip was to gain more knowledge on airline catering in relation to tourism.
+
+              Mr. Justin the Human Resource Manager gave a brief introduction about all that is done at LSG Skychefs.
+      `,
+      },
+      {
+        id: uuidv4(),
+        postDate: new Date("Wed Jun 15 2022 00:00:00 GMT+0100"),
+        isTopPost: false,
+        postsImg: careerDrive1,
+
+        postHeading: "CAREER DRIVES",
+        coverImage: careerDrive2,
+        postsText: `Career Drives are a yearly feature at Wavecrest College and serve as an opportunity to achieve our mission of creating awareness of the Hospitality Career. The Drives took place in several batches both within the College and in various schools in Lagos.
+
+              In 2023, we had the pleasure to host students of the following schools:
+
+               - Estate Senior Grammar School (22 Students)
+               - Marywood Girls College (150 Students)
+               - Alimosho Grammar School (100 Students)
+               - Wesley Girls College (15 Students Students)
+               - New Estate Baptist School (60 Students)
+               - Baptist Group of Schools (120 Students)
+               - New Estate Baptist School (25 Students)
+               - Kays Schools (25 Students)
+               - LIHS School Girls Club (15 Students)
+
+              The Hospitality Career talks highlight core areas of the Hospitality Industry as a dynamic and diverse sector, offering a wide range of career opportunities for individuals who are passionate about serving others.
+
+              Our social media handles are a major tool to reach out to many more people.`,
+      },
+      {
+        id: uuidv4(),
+        postDate: new Date("Wed Jun 15 2022 00:00:00 GMT+0100"),
+        isTopPost: false,
+        postsImg: NBTEdirector,
+        postHeading:
+          "Minutes of Meeting with Engr. Bode Daniel Majiyagbe – Director, South-West Zonal Office, NBTE and the Management of Wavecrest College of Hospitality",
+        postsText:
+          "Engr. Bode Daniel Majiyagbe had a brief meeting with the Management Team of Wavecrest College of Hospitality on Thursday 20th Janaury 2022 at 9.30 a.m. The aim of the meeting was to formally inform the College of the decision reached by the Executive Secretary Prof. Idris M. Bugaje and the Management of NBTE to establish zonal offices for NBTE in the geopolitical zones of the country. <br /><br />He explained clearly that the aim of this new development is to bring the services of the Board closer to the Stake Holders and Institutions under the NBTE and to introduce the TVET (Technical and Vocational Education and Training) initiative of the Board to all Stake Holders concerned.<br /><br />He reiterated that the main agenda for the Executive Secretary and his team is ‘Skills, Skills, Skills’ stating that accreditation of some Institutions may be put on hold until they adopt this skills frame work into their curriculum and have a clear plan of incorporating both their students, indigenes and youth of their immediate<br /><br /> Engr. Bode Daniel Majiyagbe presenting copies of the Operational Manual for National Skills Qualification Framework and ‘Skills Not Degree’ publication of the NBTE. to our Provost. <br /><br />Finally he announced that the South-West Zonal Office is located in Lagos State at Yaba College of Education which is central and easily accessible.<br /> The Provost of Wavecrest College of Hospitality – Rosana Forsuelo responded enthusiastically, stating that she and the Management Team are very happy with this new development as it is a clear and concrete sign that the NBTE is truly concerned about serving the Institutions under them better each time.<br /><br />‘Wavecrest College is also always at the service of NBTE being very involved in and promoting the NSQF’ – Rosana Forsuelo. She also mentioned the fact that we already have an office dedicated to the NSQ in Wavecrest College.<br />The meeting was brief and to the point. It ended with Engr. Bode Daniel Majiyagbe presenting copies of the Operational Manual for National Skills Qualification Framework and ‘Skills Not Degree’ publication of the NBTE. <br /><br /><br />Secretary  <br />Veronica Esode<br /> Registrar Wavecrest College of Hospitality",
+      },
+      {
+        id: uuidv4(),
+        postDate: new Date("Wed Jun 15 2022 00:00:00 GMT+0100"),
+        isTopPost: false,
+        postsImg: suicideEvent,
+        postHeading: "World Suicide Prevention Day",
+        postsText:
+          "There is growing awareness of suicide as a major public health problem, even though there is a taboo in many societies against discussing openly. Worldwide Suicide rates have increased by 60% over the last 50 years, and the increase has been particularly marked in developing countries. Suicide occurs all over the world and can take place at almost any age. <br /><br />Globally, suicide rates are highest in people aged 60 years and over. In some countries, however, the highest rates are found among the young. Notably, suicide is now among the three leading global causes of death among young people aged 15-34 years, the majority of suicides are reported in adults and older adults (60 years and older). <br /><br />Pesticide poisoning, hanging and firearms are among the most common methods of suicide globally.<br /><br />In Nigeria, Suicide is becoming a looming Public health problem and should be taken seriously.We remember popular cases that were reported of a man, a medical doctor who drove down to the third mainland bridge, parked his car, jumped into the lagoon, and died in 2019.<br /><br />-A young man identified as Kelvin committed suicide at Modinatu Street in the Ogba area of Lagos.<br />-The lifeless body of Kevin who works with a new generation bank was found dangling from a ceiling fan, in his apartment by neighbors.<br />-Young Nigerian, creative designer and founder of Uhuru Designz, the United States of America-based clothing company, Izuchukwu Madubueze reportedly committed suicide Over Fake Rape Allegations that took place on Twitter in July 2020. He even announced his own death on twitter before his death.Sadly as we know these are not all the cases and some have not been covered by the media. <br /><br />Suicides are Preventable!<br /><br />“No matter where a country currently stands in suicide prevention”, said Dr Alexandra Fleischmann, Scientist in the Department of Mental Health and Substance Abuse at WHO, “effective measures can be taken, even just starting at local level and on a small-scale”<br /><br /><ol><li>Always Check up on friends and family.</li><li>Encouraging friends and families to always take mental health seriously and to see a counselor when it’s necessary.</li><li>Responsible reporting of suicide in the media, such as avoiding language that sensationalizes suicide and avoiding explicit description of methods used, and early identification and management of mental and substance use disorders in communities and by health workers in particular</li>.<li>Social media has come to stay and some people are living fake lives. Don’t compare your life with others because you don’t know what they are going through and you may just be better off than them.</li><li>Follow-up care by health workers through regular contact, including by phone or home visits, for people who have attempted suicide, together with provision of community support, are essential, because people who have already attempted suicide are at the greatest risk of trying again.</li><li>Regular exercise, improve healthy lifestyle.</li><li>Reducing access to means of suicide is one way to reduce deaths.</li></ol> <br /><br />Below is a few list of some Mental health Organizations we have:<br /><br /><ul><li>African Mental Health Awareness & Care Initiative, Maiduguri</li><li>Mental Health Foundation, Nigeria</li><li>Gede Foundation, Abuja </li><li>UNIC Ikoyi</ul> <br /> <br /> Source: WHO, Vanguard <br /><br />  Article written by Sheba Asuquo",
+      },
+      {
+        id: uuidv4(),
+        postDate: new Date("Wed Jun 15 2022 00:00:00 GMT+0100"),
+        isTopPost: false,
+        postsImg: blogPost,
+        postHeading:
+          "‘The icing on the cake”…Wavecrest College of Hospitality Partners with Fly Africa Initiative to break Guinness World Record 2020",
+        postsText:
+          "Featuring in the <em>Guinness World Records</em> is obviously not a yearly event. This is why Wavecrest College was excited to be a part of an amazing project partnering with the Fly Africa Initiative to make an official attempt to break the Guinness World Record to make the largest cupcake mosaic logo of 60,000 cupcakes to celebrate Nigeria’s 60th Independence Anniversary in Lagos.<br /><br />For those who may not know, Fly Africa is an initiative of Outori Limited created to inspire and educate Africans through live experiences, to rise up to their great potentials by taking actions that will move the African continent forward. Mr. Wole Olagundoye the Managing Director of Outori and the brain behind the ‘Sweet 60’ project visited Wavecrest College with his Team to conclude plans and logistics on how the tail end of the project will be executed. This last lap was indeed very crucial and needed to be handled by professionals and like minds who saw the vision of the project through his own eyes. <br /><br />Choosing Wavecrest College; the place to <em>‘Discover Hospitality’ </em>to partner in the project was indeed the best decision. Wavecrest College is the First Private Monotechnic in Nigeria specialising in Hospitality careers, with over 40 years of hospitality experience built on our solid core values of Teamwork, Respect for others, Industriousness, Personal responsibility, Integrity and Spirit of service.<br /><br />The role of Wavecrest was mind-blowing – to prepare the butter icing, roll and cut out fondant icing of different colours needed to cover 60,000 cupcakes. 60,000 not 6,000! This indeed was a challenge for us but knowing the goal, we were ready to go any length in order to make it happen.<br /><br />The College became a factory overnight with all hands on deck from the usual catering staff, to our students who joyfully volunteered. Even the Provost had her own quota of fondants to cut. It was another opportunity to witness our values come alive in real time as every member of staff took up the project like it was their personal business. It took 4 days working round the clock, swimming in shades of green, grey and white fondant, kneading and cutting 3 inches each of fondant, no more no less.<br /><br /> <em>‘Thank you for everything your team did to make this historical event a success. Your warm hospitality was second to none, your team members were awesome and dedicated and I am indeed proud to be associated with you.’</em> This feedback from one of the Project Coordinators says it all!<br/><br/> At about 11 p.m. Wednesday night – 30th September, 2020 -, the gigantic-sized cooling van containing the last batch of meticulously labelled cake boxes especially those for the shimmering diamond design in the centre of the logo left Wavecrest College premises to Teslim Balogun Stadium. At that point, we could heave a sigh of relief because we knew we had achieved.<br/><br/> Wavecrest was part of ‘the icing on the cake’ the beauty that met the eye, for the ‘Sweet 60’ project, the Guinness world record breaking mosaic representing the beautiful and breath-taking 60th Anniversary logo of Nigeria’s 60th independence anniversary. <br/><br/>We would like to appreciate Mr. Wole Olagundoye for giving Wavecrest College the opportunity to be a major part of the project and the penultimate site for the last lap of preparations. A man who can embark on a project like this in spite of the challenges of 2020 is obviously a person to reckon with. His resilience drive and optimism are second to none. Fly Africa has certainly lived up to its aim of showcasing African talent and solidarity as all parties involved played their role excellently well. The major sponsors and collaborators – Lagos State Government, Dangote Sugar, Graceco Foods -Baker’s Choice and Golden Penny could not have been more supportive.<br/><br/>The other partners were a pleasure to work with here at Wavecrest College – Atinuke Soje of Nukkies Boutique cakes, Mr. Ajayi of U & I cakes, Mrs. Toyin Naiwo of EventsPedia and indeed all the amazing volunteers. With this display of team spirit, we are very hopeful that our dear country Nigeria has a brighter future. <br /><br />‘Cheers to Sweet 60’. <br /><br />Baptista Uwandu Wavecrest College Marketing Personnel, Mrs. Toyin Naiwo EventsPedia, Mrs. Helen Eboigbe <br /> – Wavecrest College Development Committee.",
+      },
+      {
+        id: uuidv4(),
+        postDate: new Date("Wed Jun 15 2022 00:00:00 GMT+0100"),
+        isTopPost: false,
+        postsImg: nbte,
+        postHeading: "NSQ PROGRAMMES AVAILABLE",
+        postsText:
+          "There are lots of different NSQs to choose from, which are regularly being reviewed and updated. They are available in five sectors i.e hospitality and tourism, power/energy/engineering, building construction, service and agro-processing as giving in the table below;<br/><br/> <b>The NSQ system – Who does/do what</b> <br /><br/><b> Sector skills councils:</b> identify, define and update employment-based standards of competence for agreed occupations. These people effectively represent the interests of the trade or industry concerned, to which the particular NSQ relates. Sector Bodies define what the job entails. For government led sector skills council, the composition should be 70 % industry/ employers and 30% trade unions, professional associations, academia etc.<br /><br /> <b>Awarding bodies:</b> design assessment and quality assurance systems, and gain sector bodies endorsement prior to submission to NBTE for accreditation of the qualification. Awarding bodies approve training and assessment centres to offer NSQs. implement and assure quality of the NSQs . Awarding Bodies are basically the top-level training certification organisations (the ones whose names and logos are on the certificates). They don’t deliver the training, they design the structures of the qualifications and accredit the certification, which the training organisations and Assessment Centres use and deliver.<br /><br /> Awarding Bodies engage with client organisations if they are large enough to have their own NSQ training and delivery departments. Awarding bodies engage with Sector Bodies and training organisations. Training organisations are effectively distributors of the Awarding Bodies’ qualifications systems.<br /><br /> They undertake external verification to ensure that candidates are being assessed fairly and consistently across all centres.<br /><br /><b> Regulatory Body: NBTE</b> accredits proposals for qualifications submitted by awarding bodies, and monitors Awarding bodies offering NSQs. NBTE effectively manages the whole system on behalf of the Government. <br /><br /> <b>Assessment centres:</b> assess NSQs according to Awarding body criteria.<br /><br /><b>Training providers: </b> deliver NSQ training in partnership with the industry, register candidates, assess and guide candidates towards achieving NSQs.",
+      },
+    ],
+  },
+
+  {
+    category: "recipe",
+    posts: [
+      {
+        id: uuidv4(),
+        postDate: new Date("Wed Jun 15 2022 00:00:00 GMT+0100"),
+        isTopPost: true,
+        postsImg: cookFish,
+        postHeading: "The Different Ways To Cook Fish",
+        postsText:
+          "Yes, you know that fish is one of the healthiest and common proteins in the world after eggs! It’s good for the heart because of the abundance of omega-3. It has other health benefits such as stroke reduction, contribution to brain health, and prevention of Alzheimer’s amongst others. <br /> <br />What you may not know are the different ways to cook fish. smiles. <br /><br />Christmas is by the corner and you may want to take the time to enjoy yourself in different ways. <br /><br />Now, follow me as I show you 5 ways to cook fish.<br /><br /> <ol><li><b>Grilling Fish:</b> When grilling fish, keep a close watch. Fish only takes a few minutes per side to cook.<br /><br /></li>  <li><b>Poaching Fish:</b> poaching refers to the cooking of fish with liquid. It might be water, stock, milk, or wine. To poach, add enough water in a pot or pan to just barely cover the fish.</li><br /><li><b>Steaming Fish:  </b>This refers to the use of a steamer basket to steam the fish. Here’s how to steam fish. <br /> <br /> <ul><li>Marinate the fish with spices, ginger, garlic, and chili peppers to make it tasty. </li><li>Pour about 2 cups of water into the pan or pot.</li><li> Use a folding steamer basket with enough room for each piece of fish to lie flat.</li><li> Place the steamer over the water, cover the pot, and leave the water to boil. </li><li>Then check to see when it’s done.</li></ul></li> <br/><li><b>Baking Fish:</b> Baking fish allows you to get the satisfying crunch of fried fish without all the fat.</li><br /> <li> <b>Deep Frying </b></li> <br /><li><b> Broiling Fish</b></li> </ol> <br />There are many ways that you can prepare fish, but some cooking methods can bring out the full flavor of fish. Want to learn more, register for our culinary program on <a href='https://wavecrest.edu.ng'> wavecrest.edu.ng.</a>",
+      },
+      {
+        id: uuidv4(),
+        postDate: new Date("Wed Jun 15 2022 00:00:00 GMT+0100"),
+        isTopPost: true,
+        postsImg: catfish,
+        postHeading: "CATFISH PEPPER SOUP",
+        postsText:
+          "INGREDIENTS: 1.2 kg fresh catfish, 110g onions, chopped (1 medium), 2 seasoning cubes, 1 tablespoons ground crayfish, 2 scotch bonnet, salt, 2 big ehuru, 3 small ground nutmeg, 1/2 teaspoon ground uziza, 2 sticks uda, Utazi leaves (bitters). METHOD Toast and grind ehuru, hit uda to just break it open. Wash catfish slices well with lime to get rid of the sliminess. Pour hot water over the fish slices and leave to stand for 1-2 minutes, drain and rinse in cold water.Place in a pot and add just enough water to cover the fish. Place over medium heat. Add onions, seasoning cube, crayfish, peppers and sprinkle some salt. Cover and bring to a boil, add ground ehuru, uziza and uda, gently shake and swirl your pot, cook for 5-8 minutes. Add about 5 leaves of shredded utazi and some water. Taste and correct seasoning. Cook for another 5-8 minutes until fish is well cooked. Take off the heat. Serve your Catfish pepper soup with boiled yam, plantain, rice, agidi (eko, corn paste pudding) etc. NOTE: You can use less scotch bonnet and add a little Cameroon pepper, it tastes really good. Instead of the different spices used here, you can simply use store-bought Pepper soup spice mix or any mixture of spices you like.",
+      },
+      {
+        id: uuidv4(),
+        postDate: new Date("Wed Jun 15 2022 00:00:00 GMT+0100"),
+        isTopPost: true,
+        postsImg: ekpangNkukwo,
+        postHeading: "EKPANG NKUKWO",
+        postsText:
+          "Ekpang Nkukwo is a traditionnal Nigetrian dish native to the Efiks and Ibibios. It is made of freshly grated cocoyam, then wrapped in leafy greens and slowly simmered with various types of fresh o smoked fish, meat, crayfish, red oul and sther spices depending on the cook's preference until cooked to perfection <br /> <br /> <b>INGREDIENTS:</b> <br /><br /><ul><li> 4 pounds cocoyam,</li><li> 4-5 bunch leafy greens cocoyam leaves,</li><li> Collard greens,</li> <li>Spinach,</li> <li>Romaine lettuce,</li><li> ½ pound boiled beef,</li><li>1-2 cups dried fish,</li><li> ½ cup ground crayfish,</li><li> ½ medium onion chopped,</li><li> 1½ -2 cups palm oil /red oil,</li><li> Bouillon cubes to taste, </li><li>1 tablespoon ground onions (optional),</li><li>1 teaspoon ground pepper,</li><li> 1 scotch bonnet pepper or any hot pepper optional, </li><li>Salt and pepper to taste.</ul><br /> <b>PROCEDURE:</b> <br /> <ul><li>Cut the beef into bite-size chunks.</li><li> Season with salt, Maggi, and onions. <li>Cook until tender.(make sure you have at least 2 to 3 cups of stock from the beef to use in cooking this dish).</li><li>Peel Coco yam and cut into large chunks and wash immediately to prevent discoloration and leave them in cool water until ready to use.</li><li>Pound the cocoyam in a mortal until there is a puree or use a grater to grate the cocoyam.  </li><li>Lightly salt and set aside.Wash leaves and tear them into medium pieces, making sure there are no holes in the leaves.</li><li>Pour about ½ cup of oil into a large pot. Set aside.</li><li> Scoop a tablespoon or so of grated cocoyam mixture onto a leaf and wrap tightly (to form the shape of fingers) making sure the cocoyam is not sticking out.</li><li>Arrange wrapped cocoyam fingers in the oil pot in a circle leaving a hole in the middle. </li><li>Repeat until the cocoyam mixture has been completely used up.</li><li>In a separate pot bring about 7 cups of water to a boil.</li><li> Add chopped onions, crayfish, fish, meat, salt, pepper, bouillon cubes, palm oil, and onions to the pot.</li><li>Let it simmer for about 10 minutes, there should be no stirring at this point until most of the ekpang firms up- add about 1/2 cup of beef stock to prevent any burns.</li><li>Then after ekpang firms up add about all the beef stock if you have any with about 3 cups water or just use 4-6 cups of hot water or more to the pot and let it cook on medium heat for about an hour until fully cooked.</li><li> You may add more water to prevent burns.</li><li>Stirring should be kept at a minimum- to prevent the ekpang from becoming mushy. </li><li>Adjust for ekpang consistency with water, salt and pepper.</li></ul>",
+      },
+    ],
+  },
+
+  {
+    category: "testimonial",
+    posts: [
+      {
+        id: uuidv4(),
+        postDate: new Date("Wed Jun 15 2022 00:00:00 GMT+0100"),
+        isTopPost: true,
+        postsImg: Udoka,
+        postHeading:
+          "I Wanted to Study in A Reputable Culinary School That is Why I Came To Wavecrest College- Udoka John",
+        postsText:
+          "<b>Hi Udoka John, you look excited to be here. What influenced your choice of training at Wavecrest college? </b><br /><br />Seeing other Chefs on television makes me so excited so I said to myself ‘I must study in a reputable place’, that’s why I came to Wavecrest College of Hospitality.<br /><br /><b>How has this training in Wavecrest equipped you for the future?</b> <br /><br />The training has added so many recipes to my cooking. I have learnt a lot of delicacies that would help me in my future business, in my home and my marriage. <br /><br/> <b>Wow, that’s beautiful. The family gets to benefit also. What’s your next line of action?</b><br/><br/>I would like to work for an organisation first to gain more experience in the industry, thereafter I will open my own eatery. <br /><br /><b>Where do you see yourself in the future, the next 5 years from now? </b><br /><br/>In the next 5 years, I want to see myself owning a BIG  eatery that people would be trooping in to get their stuff.<br /><br /><b> So be it. Congratulations to you</b>",
+      },
+      {
+        id: uuidv4(),
+        postDate: new Date("Wed Jun 15 2022 00:00:00 GMT+0100"),
+        isTopPost: true,
+        postsImg: blogPost2,
+        postHeading:
+          "“I Want to Own My Restaurant That is Why I Came to Wavecrest College for Training” -Victoria Joseph",
+        postsText:
+          "<b>Hello, may we meet you?</b> <br /><br /> My name is Victoria Joseph <br /><br /> <b>What inspired you to study at Wavecrest College?</b> <br /><br />  I have always wanted to own my restaurant so looking for a school that can give me the basic knowledge, skills and also the right ingredients combined to produce a mouth-watering delicacy. <br /><br /> I searched through the internet and I came across Wavecrest College. I found out that it is a reputable institution and I decided to give it a shot. I came to visit and I saw the environment, everybody was well organized  so I said this is the perfect place for me.  <br /><br /><b> How has this training in Wavecrest equipped you for the future? </b> <br /><br /> The training has equipped me to create a lot of recipes now and the Chef filled us with so much knowledge, little secrets and tips I never knew. I am sure that when I open my restaurant is going to be a hit.<br /><br /><b> So you intend to open a restaurant? </b> <br /><br />Yes… that’s why I am here. I want to own my own restaurants with a lot of branches in Lagos, and… everywhere in the world. <br /><br /><b> Where do you see yourself in the future, the next 5 years from now? </b> <br /><br /> I see myself owning my restaurant and having lots of outlets around the globe.. <br /><br /> Thank you for your time",
+      },
+    ],
+  },
+
+  {
+    category: "hospitality",
+    posts: [
+      {
+        id: uuidv4(),
+        postDate: new Date("Wed Jun 15 2022 00:00:00 GMT+0100"),
+        isTopPost: true,
+        postsImg: airport,
+        postHeading:
+          "Airline Travel: Indian Domestic Flights Resumes After Covid-19 Shutdown",
+        postsText:
+          "Good news for the Hospitality and Tourism Industry: Indian domestic flights have opened up today after about 2 months of Covid-19 lock-down. <br /><br /> This is good news for the world of Hospitality and Tourism because India is the second-most populous country in the world with over 1 billion people. While major parts of the Hospitality Industry remain closed, we are excited that countries are beginning to open up. <br /><br />Remember to stay safe and observe all health guidelines.",
+      },
+      {
+        id: uuidv4(),
+        postDate: new Date("Wed Jun 15 2022 00:00:00 GMT+0100"),
+        isTopPost: true,
+        postsImg: airTravels,
+        postHeading: "The future of Air Travels after COVID 19- Sheba Asuquo",
+        postsText:
+          "It is no longer news that Aviation is one of the sectors most affected by the COVID-19 pandemic. <br /><br />As the lockdown continues to ease off in most parts of the world a lot of people are still wondering what travel would be like. <br /><br />Research from The International Air Transport Association (IATA), shows that just 14% of consumers say they’ll resume travel as soon as restrictions are lifted.<br /><br />Nevertheless, the airports and airlines are working round the clock and putting measures in place to ensure safe travel post COVID-19.<br /><br /> You should expect to see the following in the airports and airlines as we look to a brighter future.<br /><br /> Case study: EMIRATES AIRLINE <br /><br /><ol><li>Health screening protocols will be carried out before every journey to ensure that passengers are well enough to fly. </li><li>Rapid COVID-19 Passenger testing for specific destinations. </li><li>Enhanced cabin cleaning and thermal screening for fever detection. </li><li>Microbial shields at check-in and social distancing measures.</li><li> Complimentary travel hygiene kits for all passengers such as face masks, gloves, anti-bacterial wipes and hand sanitizer for extra protection during the journey.</li><li> Staggered boarding in small groups.</li><li> Modified in-flight service for health and safety reasons. </li><li> Dedicated cabin service attendant for onboard hygiene</ol><br /> Even though some experts say that it may take up to 2 years for the aviation industry to be as it was before the COVID-19, we hope that it will be quicker than imagined.",
+      },
+      {
+        id: uuidv4(),
+        postDate: new Date("Wed Jun 15 2022 00:00:00 GMT+0100"),
+        isTopPost: true,
+        postsImg: hotel,
+        postHeading:
+          "HOTEL INDUSTRY RECOVERY: VALUE RETENTION AND SUSTAINABILITY",
+        postsText:
+          "<b>Written by Alec Hills; Article from</b> <a href='https://www.tourism-review.com/.'>https://www.tourism-review.com/. </a> <br /><br /> <em>After the drop in demand and the massive loss of control in the past months, the hotel industry is longing for the end of the Coronavirus pandemic. At the operational level, intensive hygiene concepts and process redesigns for a successful reopening are being implemented and questions are being asked about when everything will return to ‘normal’.</em> <br /><br /> <b> Bright Spots but Far from Stable.</b><br /><br /> However, people in the hospitality industry are slowly becoming aware that the scale of this crisis is so drastic that the future has already changed direction. <b>Thus, for the hotel industry, it is now a matter of making a fresh start and integrating socio-economic challenges into their own business model. </b><br />The first bright spots are slowly emerging in the travel market. However, it will still take some time to stabilize. Thanks to improving numbers of infections and the enormous importance of tourism for the national economies, hotels in most European countries are already open again. <br /><br />With the reopening of the hotel and travel markets, it is now up to hoteliers to cope with the effects of the crisis.<br /><br /> <b> Many Questions Arise</b> <br /><br /> Hotel operators are facing a challenge of whether the reopening will pay off at all. If so, exactly when and under what legal conditions, without losing any more money? Can the lease be paid again?<br /><br /> Investors and asset managers, in turn, worry about the tenants securing cash flows and the commitments and promises made to investors. Active buyers are wondering which markets, segments and concepts are the fastest to emerge from the crisis mode and offer opportunities for increasing value.<br /><br /> <b>Localization of a Globalized World.</b><br /><br /> A trend that is already emerging is the localization of a globalized world. The hotel industry recovery can benefit from local productions booming and networks being localized.<b>In the future, the viability of the business relationship between landlord, tenant, lender and borrower, client and customer will be put to the test.</b>  Hotel owners, operators and banks have to secure values and be warry of the sustainability of business models.<br /><br /> The world as we know it is currently dissolving, at least in some areas. At the same time, however, there are new long-term perspectives. In addition to financial aid and fair burden-sharing, it is now a matter of developing business models with the involvement of all stakeholders. Unilateral solutions at the expense of a partner will quickly lead to major problems in the next few years.",
+      },
+      {
+        id: uuidv4(),
+        postDate: new Date("Wed Jun 15 2022 00:00:00 GMT+0100"),
+        isTopPost: true,
+        postsImg: decorating,
+        postHeading: "Top 3 Creative Ideas for Decorating Your Dining Table",
+        postsText:
+          "Having a small party at home is also an opportunity to display your entertainment skills. Since the meal is the main focus of the entire celebration, you should choose a table centerpiece that will really impress your guests but you don’t have to break the bank to get it done. Here are some creative ideas for a stylish centerpiece.<br /><br /><ol> <li>Candles: Place a large candle in a glass container surrounded by stones and shells.  Nothing makes a table look more elegant than soft, glowing candles.</li><img src='../assets/imgs/blog-candle.jpeg' alt='candles' /> <br /><br /><li>Lemons, limes, and oranges add bright pops of color to a table without breaking the budget. Stack them in tall, narrow vases or fill round, wide bowls with them. </li><br /><br /><li>Wine Bottles: Take a few bottles from your previous party and remove the labels. Then fill them with thin branches of flowers and tie them with a rope or ribbons. These types of arrangements work really well at picnics and outdoor gatherings.</li></ol>",
+      },
+    ],
+  },
+
+  {
+    category: "others",
+    posts: [
+      {
+        id: uuidv4(),
+        postDate: new Date("Wed Jun 15 2022 00:00:00 GMT+0100"),
+        isTopPost: true,
+        postsImg: developingSkills,
+        postHeading:
+          "Developing Skills for a Better Life – Vocational Training upgraded",
+        postsText:
+          "<em>Veronica Esode</em><br /><em> Registrar </em><br /><em>Wavecrest College of Hospitality </em> <br /><br />In the world today, the exponential growth rate of graduate unemployment has become a serious cause of anxiety. It is a challenge that most economies are facing under the current economic circumstances (Oppong & Sachs, 2015). <br /><br />More so, with the advent of COVID-19, the world we have always known is not just rapidly changing, NO! In fact, we can hardly recognize it, and yet we must survive and thrive with whatever we can lay our hands-on. <br /><br />Some very well-known causes of unemployment in Nigeria include an excessive supply of university graduates as compared with available job opportunities, insufficient hands-on and practical training to match job requirements, and a non-existent curriculum for training in Life Skills which is of utmost importance for survival and thriving in the work environment.<br /> <a href =' https://www.nepad.org/skillsportalforyouth/good-practice/depth-hospitality-and-life-skills-training-guarantees-girl-child. '> https://www.nepad.org/skillsportalforyouth/good-practice/depth-hospitality-and-life-skills-training-guarantees-girl-child. </a> <br /><br />The figures from JAMB 2019 Admissions speaks volumes. For all courses apart from Education and Agriculture, applicants are at least 3 times more than the available spaces in Higher Institutions. In the case of Medicine and Law, they are up to 15 times more than the available spaces.<br /><br />However, for Polytechnics and Monotechnics who have a quota of 17, 574, only 4,982 were admitted. That is 72%  of spaces are waiting to be filled. This may be due to the widespread notion that the University Degree is superior to the Diplomas from other Higher Institutions. It could also be due to the fact of the poor standards of some Institutions. Nevertheless, times have changed and we need to look forward and ahead.<br /><br /> Conrad Burns said ‘Vocational education programs have made a real difference in the lives of countless young people nationwide; they build self-confidence and leadership skills by allowing students to utilize their unique gifts and talents’.<br /><br /> The Nigerian government is doing its bit to establish policies and strategies to salvage the situation. Technical and Vocational Training (TVET) System <a href='https://unevoc.unesco.org/wtdb/worldtvetdatabase_nga_en.pdf'>https://unevoc.unesco.org/wtdb/worldtvetdatabase_nga_en.pdf</a> and the National Skills Qualification Framework (NSQF) for example aim to assist the federal and state education authorities in their effort to revitalize, reform and expand the provision of skills, vocations, science, and technology to meet the nation’s present and future socio-economic needs.<br /><br /> Nevertheless, as individuals, we need to wake up to the reality on the ground. Vocational training and Technical skills are extremely pertinent today. Truth be told,  the professional courses are needed – we always need Doctors, Engineers, and Lawyers but as we see from the JAMB statistics, these courses are highly saturated on the level of admissions and much more so on the level of employment. After the struggle to gain admission, getting a job is even more difficult. <br /><br />Why do we keep complicating our lives when there are other options and possibilities, lucrative and mind-blowing career paths that Nigerian youths can follow? Many applicants are still out there seeking admission into ‘professional’ courses. This mentality among Parents, Guardians, and the Applicants themselves needs to be transformed and enlightened if the scourge of unemployment is to be eradicated. ",
+      },
+      {
+        id: uuidv4(),
+        postDate: new Date("Wed Jun 15 2022 00:00:00 GMT+0100"),
+        isTopPost: true,
+        postsImg: uchyfide,
+        postHeading:
+          "The Three E’s of Wavecrest College of hospitality by @Uchyfide",
+        postsText:
+          "<b>Education:</b> I have been a lover of good and wholesome meals before I attended Wavecrest College. But Wavecrest took my brain, my ideas about food and hospitality, and in fact my whole body system back to factory settings. In a few words, I have gained basic and standard knowledge about good cooking skills and accurate measurement in the kitchen, personal hygiene, professional ethics, and morals. <br /><br /><b>Empowerment:</b> I have been empowered with a heart of a lioness to break limits and do exploits, to stand out and be unapologetically independent. I now have self-confidence, honesty, truth, and godly morals.<br /><br /><b>Employment: </b> I have employed myself full-time. I am into cooking, serving, and educating others at UCHYFIDE CULINARY. It didn’t come that easy, with much work, dedication, and most importantly with the help of God Almighty, I am a full-time staff and manager of Uchyfide CULINARY. I AM BREAKING LIMITS. THANK YOU WAVECREST COLLEGE OF HOSPITALITY.",
+      },
+      {
+        id: uuidv4(),
+        postDate: new Date("Wed Jun 15 2022 00:00:00 GMT+0100"),
+        postsImg: studentLetter,
+        isTopPost: true,
+        postHeading: "A Letter to my students during lockdown",
+        postsText:
+          "Heyyy!!!<br /> <br /> Who screamed? My house! The windows, the nets, the doors, and the floors everything is screaming…why? They have never experienced such ‘squeaky clean’ scrubbing ever. Everywhere … What is happening? For the past 3 weeks everyone has been at home. Something in town…Corona Virus! Run! <br /><br /> The sink: <em>‘I am constantly wet’.</em> (Everyone is washing their hands). I can see one person about 15 times in a day. My assistant (liquid soap) does not last anymore for up to 2 days.<br /><br /> The Mirror: I am not lacking faces anymore. (Those washing their hands have been admiring themselves)<br /><br /> The Door: The elbow has become my new love. (Everyone tries to open with the elbow.) <br /><br />The floor: I get to be swept and scrubbed daily. My new perfume is bleach! <br /><br />My dear students, <br /><br />The situation we find ourselves in, has been highly unprecedented. We are shocked, the whole world has been taken unawares. And we find that in such situations, we don’t know how to react. There are many ways one can react: ‘fear, learn or grow…’ <br /><br /> We have chosen to act, we want to… <br /><br />At this period too, the fear of the effects of the pandemic, including reduced supply of our basic needs for the day make it more worrisome.  <br /><br />But what do we do? Where do we go from here?<br /><br /> It is very tough for you and for most of us. And now, we need to do our classes online! Technology use comes with its varying challenges: not having adequate mobile devices to download the appropriate app, insufficient power supply with reduced battery power of our devices, and the increasing cost of data! Name it!….<br /><br /> Can we make the most of our situation? We decided at the point of the lockdown that we would start classes online if ‘stay home’ continues. We agreed to use the most accessible means for many of us for classroom teaching and learning. However, as the day goes by we realise that the challenges get rougher. And we don’t know, what can we do?<br /><br /> While we work and pray for the best…Do you know you can save your data and your battery life? Here are a few tricks I use:<br /><br /> Keep a timetable and keep busy in the house. Have time to clean, study, be on the phone, watch movie, go to bed etc. When you engage in many activities, you would not be ‘hooked’ on your phone.Switch off phones when not in use for a long time. For example when going to bed. <br /><br /><ul><li>Put your phone settings so that it does not automatically download media. You download the ones you believe are of interest to you. As you know, your contacts send many things on Whatsapp and some are repeated in groups.</li><li> Set your Whatsapp messages to ‘silent’, your phone will not beep all the time message comes in.</li><li> Switch on your data at a fixed time when you plan to be online or to check and reply messages.</li><li> If you are working with a laptop, try to work offline most of the time and then copy and paste your work when you are ready to reply or send email for example. Switch off apps that run or update automatically once you are connected to internet.</li><li> You may want to exit from Whatsapp groups that do not add much value to you. Evade clicking on all messages on the group especially if you are aware that the sender usually sends unimportant messages. </li><li>Limit sending videos and making video calls. Use messages to fix appointment for when you want to call or to be called.</li><li> Plan your social media presence especially on Facebook.<b>Must I like all selfie photos… </b> NO! </ul><br /> <br /> Back to our letter… <br /> <br /> What can I learn today outside school work? I won’t go to bed unless I learn something ‘new’…’ Let’s grow on the inside…’ like St. Josemaria advises. ‘If I make it out of this period what account do I give to myself, to others and God of making use of the talents I have…’ Creativity, initiative within the limits of the so little I have left…. This is part of Hospitality… remember…Wavecrest! …Discover Hospitality… <br />What can I do for someone today? <br /><br /><em> ‘I will pray and see how…’ </em>Yes, that’s the spirit… <br /><br />We can …God help us! 😀 <br /><br /> Nwabachili, 16-04-2020 <br /><br />p.s. I plan to learn to use the lawnmower today as the gardener has not been at work, and the grass is almost seeming like short trees.  I am afraid ‘creepy crawly’ may soon become a friend to the compound! I would see if I send you a photo when I start using it.",
+      },
+    ],
+  },
+];
+
+export const blog = {
+  BLOGS,
+};
