+++ src/config/constants/editable-copy/home/values.ts
@@ -1,3 +1,5 @@
+import { v4 as uuidv4 } from "uuid";
+import { BlogPost } from "components/layouts/Blog/Blog";
 import { PartnersImgCard } from "components/layouts/Home/Partners";
 
 const peace1 = "/assets/imgs/peace1.jpg";
@@ -76,9 +78,13 @@ const announcements = [
   },
 ];
 
-export const CARD_POST = [
+export const CARD_POST: BlogPost[] = [
   {
-    postId: "61911947-aec1-4326-93b2-38cc1303da42",
+    id: uuidv4(),
+    isTopPost: false,
+    postDate: new Date(
+      "Tue Jul 09 2024 14:12:01 GMT+0100 (West Africa Standard Time)"
+    ),
     name: "Chinonyerem Favour",
     text: "Studying in Wavecrest has made my dreams broader; I haven't gotten the feel of any limitations ever.My colleagues and customers are always captivated by my performance and keep asking me about where I trained.",
     postHeading: "STAND YOUR GROUND  - THE EARLIER, THE BETTER",
@@ -88,7 +94,11 @@ export const CARD_POST = [
       "My name is Chinonyerem Favour, am 20 years of age. At the time I finished secondary school I couldn't get into University because I wasn't old enough so I spent one year here and there doing nothing of great importance before I took my (JAMB) – This is a National exam everyone who intends to go the University takes in Nigeria and in my opinion is really not a true test of one’s knowledge as there are several hitches one may encounter in writing this exam especially when it became a computer based exam. I took my JAMB exam twice seeking to get into a desired University for a desired course but I wasn't fortunate enough because I did not make the cut-off mark for the course. This meant that I already ‘wasted’ two years trying to get into University.After the second trial, I decided to give my true passion a chance to come alive! The truth is I have always been a lover of the fields in Hospitality especially that of culinary arts and event planning. However, in the part of the world I live, hospitality is not given enough credit for several reasons. Parents prefer their children doing more ‘prestigious’ courses like Law, Engineering and Medicine. This is true especially for children who are brilliant and brainy and unfortunately even for those who are not – there is this ‘class’ thing in the minds of parents.For me, it was not easy at first to convince my parents and the teachers in my life to allow me pursue a career in Hospitality because I was considered a brilliant and brainy student in Secondary School. But I gave it a try and succeeded. When I finally started this journey, I realized how much time I had wasted before pursuing my true dreams.  3 years after Secondary School was when I managed to convince my parents amidst different bad reviews from people telling me I was going to waste my knowledge if I pursued Hospitality courses. This time I did not bother listening to them and I set out to find the best hospitality school so I would not give anyone any opportunity to mock me later if I did not get the best. I did a Google search for the best hospitality schools in the State where I lived and I wrote down the 10 best that I saw. I started going to each of them one by one to make enquiries in person and see what they had to offer. Wavecrest was the 4th and last school I visited.",
   },
   {
-    postId: "ea9a4e7d-025f-4509-a47d-23661f7edd7a",
+    id: uuidv4(),
+    isTopPost: false,
+    postDate: new Date(
+      "Tue Jul 09 2024 14:12:01 GMT+0100 (West Africa Standard Time)"
+    ),
     name: "Udodirim Peace",
     postHeading:
       "My dreams have improved because of Wavecrest College and the exposure it has given me. ",
@@ -99,7 +109,11 @@ export const CARD_POST = [
       "My Name is Udodirim Peace and I am 28 years old. Before I came to Wavecrest I was working in a catering establishment, where I discovered my interest in hospitality, I was good with my work but I was earning very little because I had no formal training or any qualification. I had met graduates of the College and I admired them a lot. In fact, they lit the fire of Education and Excellence that kept burning in me till date. My dream was to graduate and build my career in hospitality (Room Division/ Customer Relation) and it still is. Finances have been my major challenge. But I am strongly determined to achieve my dreams. I have learnt a lot with subjects like Culinary, Research, Communication, Presentation, Simple accounting and entrepreneurial skills, Customer service where we had both theoretical and practical classes.With my training at the College, I have taught a friend of mine cake making and helped one of my classmates to improve her yoghurt making. I have also taught school girls skills on baking and cooking with proper recipes, it was a club called Home makers club at Gbaja Girls Junior Secondary School, Surulere, Lagos. Every Wednesday, Myself and the group did this for about 3 years. It helped me appreciate the fact that knowledge which is not passed on is a waste. I did my internship at Installations Limited Interior Designs and Decorations working as a Front Desk Associate and after my One-Year Internship, I was retained as a part-time staff in the Organization working on weekends and getting income to support my fees and daily upkeep at the College. My mentor  and former boss has helped me to discover my potential and my most dominating skill, yes! We worked together for 4 years even while I was schooling. Her mentorship has been nothing but a blessing. I assure you that Wavecrest is an Institution you can trust. Wavecrest has the best interest of their students at heart and will do so much to see that the students graduate.I graduated from the College last year with a Higher National Diploma in Hospitality Management. I am currently on Swiss scholarship to study in Boma International Hospitality College, Kenya. I decided to apply as a result of my interest to grow further in my hospitality career. Fortunately, I was selected among the many other students who applied.  I'm grateful to God for this great opportunity to experience different things in my career and other aspects of my life in general.Here in Boma International Hospitality College, I am doing a Certificate course  in Hospitality Operations which is mainly based on practical classes. This has really helped me in improving my skills acquired in Wavecrest College. Boma International Hospitality College is like the best thing that has happened to my career. My course consists of two (2) terms or semesters which last for about 13 weeks in each term. We study for 4 weeks and then take the first Continuous Assessment Test (CAT) and the second CAT in the 9th week and finally the end of term examinations takes place in the 13th week. The college timetable is quite flexible and yet loaded with valuable lessons. Boma International Hospitality College focuses solely on the main courses that the students require to do exploit in the industry after they graduate and get jobs. Besides, they teach according to the current trends in the hospitality industry. In other words, they do not only  rely on basic operational standards of teaching but they constantly adjust their teaching system to include what the students will meet in the industry and the school has given me the opportunity to consistently practice these skills which are  needed for me in the industry. Moreover, I have learnt both in practical and theoretical situations analysis and critical thinking, food production, food and beverage cost control and even Property Management System (PMS) for Room Division and Front Office with the software Opera PMS. In a nutshell, my experience in Boma International Hospitality College has been nothing but fulfilling. I feel like I have finally arrived and I can't wait to put everything I have learnt into practice. My dreams have not changed  but rather have been improved because of Wavecrest College and the exposure it has given me. My main focus is to grow in the industry to the highest possible position available for people as prepared and ready as I am. My appreciation to everybody and every institution that made this journey possible cannot be contained in this report, because it is too long and heavily loaded. Permit me to write a different report for it because my heart is filled with gratitude to Wavecrest College, The Swiss Government and Boma International Hospitality College.",
   },
   {
-    postId: "77e481db-29ce-4200-85a8-6eb8b8a0917c",
+    id: uuidv4(),
+    postDate: new Date(
+      "Tue Jul 09 2024 14:12:01 GMT+0100 (West Africa Standard Time)"
+    ),
+    isTopPost: false,
     name: "Dyah Thelma",
     postHeading:
       "How Catering and Deep Cleaning Became my Top Skills…Dyah Thelma",
@@ -113,7 +127,7 @@ export const CARD_POST = [
 
 export const defaultStories = CARD_POST;
 
-export const defaultPartnersLogos: PartnersImgCard[] = [
+export const partners: PartnersImgCard[] = [
   {
     companyName: "IshkTolaram",
     image: "/assets/imgs/IshkTolaram.png",
@@ -157,5 +171,5 @@ export const homeConstants = {
   discoverHospitality,
   announcements,
   defaultStories,
-  defaultPartnersLogos,
+  partners,
 };
